Moved layout logic to new file
