+++ components/tarotSpreadSelectionBox/TarotSpreadSelectionBox.tsx
@@ -1,12 +1,14 @@
+import React, { useContext, useState } from 'react';
 import { Typography, useTheme } from '@mui/material';
 import Box from '@mui/material/Box';
-import React, { useContext, useState } from 'react';
 import { TarotDeckContext } from '../../context/TarotDeckContext';
 import SingleTarotCard from '../singleTarotCard/SingleTarotCard';
+import getLayout from './getLayout';
 
 interface TarotSpreadAction {
   onClick: () => void;
 }
+
 export interface TarotSpreadSelectionProps extends TarotSpreadAction {
   title: string;
   amountOfCards?: number;
@@ -14,15 +16,6 @@ export interface TarotSpreadSelectionProps extends TarotSpreadAction {
   zIndex?: number;
   description: string;
   onClick: () => void;
-}
-
-interface CardLayout {
-  images: string[];
-  gridTemplateColumns: string;
-  gridHover: string;
-  gridVerticalOffset?: string;
-  verticalCardSpacing?: number;
-  styles?: Array<{ sx: React.CSSProperties }>;
 }
 
 const TarotSpreadSelectionBox: React.FC<TarotSpreadSelectionProps> = ({
@@ -38,98 +31,12 @@ const TarotSpreadSelectionBox: React.FC<TarotSpreadSelectionProps> = ({
   const [isHovered, setIsHovered] = useState<boolean>(false);
   const cardTransitionTime = '300ms';
   const gridTransitionTime = '120ms';
-
-  const SingleTarotCardBaseStyles: React.CSSProperties = {
-    border: '4px solid white',
-    position: 'relative',
-    width: '100px',
-    transition: `transform ${cardTransitionTime} ease-in`
-  };
-
-  const getTransformStyle = (
-    index: number,
-    centerIndex: number
-  ): React.CSSProperties => {
-    const transforms = {
-      left: 'scale(1.2) rotate(-2deg) translate(0px, -10px)',
-      right: 'scale(1.2) rotate(2deg) translate(0px, -10px)',
-      center: 'scale(1.2) translate(0px, -10px)'
-    };
-
-    if (!isHovered) return { transform: 'none' };
-
-    return {
-      transform:
-        index < centerIndex
-          ? transforms.left
-          : index > centerIndex
-            ? transforms.right
-            : transforms.center
-    };
-  };
-
-  const img = {
-    a: tarotDeckData[51].image_link,
-    b: tarotDeckData[10].image_link,
-    c: tarotDeckData[3].image_link,
-    d: tarotDeckData[1].image_link,
-    e: tarotDeckData[5].image_link,
-    f: tarotDeckData[6].image_link
-  };
-
-  const cardLayouts: Record<number, CardLayout> = {
-    1: {
-      images: [img.d],
-      gridTemplateColumns: 'auto',
-      gridHover: 'auto'
-    },
-    3: {
-      images: [img.c, img.d, img.e],
-      gridTemplateColumns: '50px 50px 100px',
-      gridHover: '80px 80px 100px',
-      gridVerticalOffset: '10px',
-      verticalCardSpacing: 10
-    },
-    4: {
-      images: [img.b, img.c, img.d, img.e],
-      gridTemplateColumns: '50px 50px 50px 100px',
-      gridHover: '80px 80px 80px 100px'
-    },
-    5: {
-      images: [img.b, img.c, img.d, img.e, img.f],
-      gridTemplateColumns: '25px 30px 30px 25px 100px',
-      gridHover: '45px 60px 60px 45px 100px',
-      gridVerticalOffset: '20px',
-      verticalCardSpacing: 10
-    },
-    7: {
-      images: [img.a, img.b, img.c, img.d, img.e, img.f, img.d],
-      gridTemplateColumns: '20px 25px 30px 30px 25px 20px 100px',
-      gridHover: '60px 70px 80px 80px 70px 60px 100px',
-      gridVerticalOffset: '40px',
-      verticalCardSpacing: 25
-    }
-  };
-
-  const generateStylesForLayout = (
-    layout: CardLayout
-  ): Array<{ sx: React.CSSProperties }> => {
-    const { images, verticalCardSpacing } = layout;
-    const centerIndex = Math.floor(images.length / 2);
-
-    return images.map((_, i) => ({
-      sx: {
-        ...SingleTarotCardBaseStyles,
-        ...getTransformStyle(i, centerIndex),
-        zIndex:
-          amountOfCards !== 4 ? images.length - Math.abs(centerIndex - i) : 1,
-        top: `${Math.abs(centerIndex - i) * verticalCardSpacing || 0}px`
-      }
-    }));
-  };
-
-  const layout = cardLayouts[amountOfCards];
-  layout.styles = generateStylesForLayout(layout);
+  const layout = getLayout(
+    tarotDeckData,
+    amountOfCards,
+    cardTransitionTime,
+    isHovered
+  );
 
   return (
     <Box
@@ -144,12 +51,8 @@ const TarotSpreadSelectionBox: React.FC<TarotSpreadSelectionProps> = ({
         transform: isHovered ? 'scale(1.025) ' : '',
         zIndex: zIndex
       }}
-      onMouseEnter={() => {
-        setIsHovered(true);
-      }}
-      onMouseLeave={() => {
-        setIsHovered(false);
-      }}
+      onMouseEnter={() => setIsHovered(true)}
+      onMouseLeave={() => setIsHovered(false)}
     >
       <Box
         sx={{
@@ -165,19 +68,17 @@ const TarotSpreadSelectionBox: React.FC<TarotSpreadSelectionProps> = ({
           display='grid'
           justifyContent='center'
           gridTemplateColumns={
-            isHovered
-              ? cardLayouts[amountOfCards].gridHover
-              : cardLayouts[amountOfCards].gridTemplateColumns
+            isHovered ? layout.gridHover : layout.gridTemplateColumns
           }
           sx={{
             position: 'relative',
             bottom: layout.gridVerticalOffset,
-            transition: `all ${cardTransitionTime}`
+            transition: `grid ${cardTransitionTime}`
           }}
         >
           {Array.from({ length: amountOfCards }).map((_, i) => {
-            const img = cardLayouts[amountOfCards].images[i];
-            const { sx } = cardLayouts[amountOfCards].styles[i];
+            const img = layout.images[i];
+            const { sx } = layout.styles[i];
 
             return (
               <SingleTarotCard
@@ -195,7 +96,7 @@ const TarotSpreadSelectionBox: React.FC<TarotSpreadSelectionProps> = ({
           textAlign='center'
           fontSize='30px'
           fontWeight={700}
-          color={palette.pinks.darker}
+          color={palette.pinks.dark}
           p={4}
         >
           {title}
@@ -205,9 +106,10 @@ const TarotSpreadSelectionBox: React.FC<TarotSpreadSelectionProps> = ({
           p={4}
           pt={0}
           sx={{
+            fontWeight: 600,
             textAlign: 'center',
             fontSize: '18px',
-            color: palette.pinks.contrastLightAAA
+            color: palette.pinks.dark
           }}
         >
           {description}
